Moved the commands into their own individual classes for easy management.
